+++ ab2d-events-client/src/main/java/gov/cms/ab2d/eventclient/clients/EventClient.java
@@ -1,0 +1,8 @@
+package gov.cms.ab2d.eventclient.clients;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import gov.cms.ab2d.eventclient.events.LoggableEvent;
+
+public interface EventClient {
+    void send(LoggableEvent requestEvent) throws JsonProcessingException;
+}
